+++ pygccxml/declarations/free_calldef.py
@@ -73,7 +73,7 @@ class free_function_t(free_calldef_t):
         free_calldef_t.__init__(self, *args, **keywords)
 
     def get_mangled_name(self):
-        if not self._mangled and not self._demangled \
+        if not self._mangled \
            and '<' not in self.name and not self.overloads:
             # it is possible we deal with C function, so lets put it name as
             # mangled one