+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Version 2.0.0 (unreleased)
 
 1. Drop support for ```GCC-XML```.
 
+  * Drop all the demangled name attributes
+
+  * Drop declarations.class_t.USE_DEMANGLED_AS_NAME
+
 2. Drop support for ```Python 2.6``` and ```Python 3.2```
 
 Version 1.9.0